+++ Demo/Empty.build.ps1
@@ -1,5 +1,0 @@
-
-<#
-.Synopsis
-	Empty script.
-#>